Fix incorrect actix scope
